refactor(client): promote ApiClientUtils to public ApiClientFactory
Renames the internal ApiClientUtils to io.unitycatalog.client's public
ApiClientFactory -- the single canonical entry point for constructing
an ApiClient -- with a convenience overload for callers with no engine
versions to report. All former ApiClientUtils callers (hadoop's
CredPropsUtil and GenericCredentialFetcher, UserAgentTest) now call the
factory; no wrapper or delegation layer remains.

The connector's engine-version detection cannot move down (it
references Spark and Delta classes the client module does not depend
on), so the old spark ApiClientFactory becomes EngineVersions, and
UCSingleCatalog passes EngineVersions.appEngineVersions() to the
promoted factory.

Phase 1 of #1181; a follow-up routes TestUtils.createApiClient through
the promoted factory.
